+++ multiqc/modules/fastqc/fastqc.py
@@ -75,7 +75,6 @@ class MultiqcModule(BaseMultiqcModule):
 
         # Filter to strip out ignored sample names
         self.fastqc_data = self.ignore_samples(self.fastqc_data)
-
         if len(self.fastqc_data) == 0:
             raise UserWarning
 
@@ -212,21 +211,17 @@ class MultiqcModule(BaseMultiqcModule):
         # Prep the data
         data = dict()
         for s_name in self.fastqc_data:
+            data[s_name] = dict()
             bs = self.fastqc_data[s_name]["basic_statistics"]
-            try:
-                # FastQC reports with 0 reads will trigger a KeyError here
-                data[s_name] = {
-                    "percent_gc": bs["%GC"],
-                    "avg_sequence_length": bs["avg_sequence_length"],
-                    "total_sequences": bs["Total Sequences"],
-                }
-            except KeyError:
+            # Samples with 0 reads and reports with some skipped sections might be missing things here
+            data[s_name]["percent_gc"] = bs.get("%GC", 0)
+            data[s_name]["avg_sequence_length"] = bs.get("avg_sequence_length", 0)
+            data[s_name]["total_sequences"] = bs.get("Total Sequences", 0)
+
+            # Log warning about zero-read samples as a courtesy
+            if data[s_name]["total_sequences"] == 0:
                 log.warning("Sample had zero reads: '{}'".format(s_name))
-                data[s_name] = {
-                    "percent_gc": 0,
-                    "avg_sequence_length": 0,
-                    "total_sequences": 0,
-                }
+
             try:
                 # Older versions of FastQC don't have this
                 data[s_name]["percent_duplicates"] = 100 - bs["total_deduplicated_percentage"]
@@ -466,7 +461,8 @@ class MultiqcModule(BaseMultiqcModule):
                     self.avg_bp_from_range(d["base"]): d for d in self.fastqc_data[s_name]["per_base_sequence_content"]
                 }
             except KeyError:
-                pass
+                # FastQC module was skipped - move on to the next sample
+                continue
 
             # Old versions of FastQC give counts instead of percentages
             for b in data[s_name]: